Fixed errors and adapted to Crossdata-0.5.0
