+++ testsAT/src/test/java/com/stratio/crossdata/testsAT/specs/catalog/CreateCatalogSpec.java
@@ -5,8 +5,8 @@ import com.stratio.crossdata.common.exceptions.ExecutionException;
 import com.stratio.crossdata.common.exceptions.ParsingException;
 import com.stratio.crossdata.common.exceptions.UnsupportedException;
 import com.stratio.crossdata.common.exceptions.ValidationException;
-import com.stratio.crossdata.specs.BaseSpec;
-import com.stratio.crossdata.specs.Common;
+import com.stratio.crossdata.testsAT.specs.BaseSpec;
+import com.stratio.crossdata.testsAT.specs.Common;
 import org.testng.Assert;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
