+++ testsAT/src/test/java/com/stratio/crossdata/testsAT/specs/clusters/AddDatastoreSpec.java
@@ -21,12 +21,11 @@ import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
 
 import com.stratio.crossdata.common.exceptions.ApiException;
-import com.stratio.crossdata.common.exceptions.ManifestException;
 import com.stratio.crossdata.common.exceptions.validation.ExistNameException;
 import com.stratio.crossdata.common.result.CommandResult;
 import com.stratio.crossdata.common.result.Result;
-import com.stratio.crossdata.specs.BaseSpec;
-import com.stratio.crossdata.specs.Common;
+import com.stratio.crossdata.testsAT.specs.BaseSpec;
+import com.stratio.crossdata.testsAT.specs.Common;
 import com.stratio.tests.utils.ThreadProperty;
 
 import cucumber.api.java.en.Given;
