+++ testsAT/src/test/java/com/stratio/crossdata/testsAT/specs/clusters/DetachClusterSpec.java
@@ -12,8 +12,8 @@ import com.stratio.crossdata.common.exceptions.ParsingException;
 import com.stratio.crossdata.common.exceptions.UnsupportedException;
 import com.stratio.crossdata.common.exceptions.ValidationException;
 import com.stratio.crossdata.common.result.CommandResult;
-import com.stratio.crossdata.specs.BaseSpec;
-import com.stratio.crossdata.specs.Common;
+import com.stratio.crossdata.testsAT.specs.BaseSpec;
+import com.stratio.crossdata.testsAT.specs.Common;
 
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
