+++ testsAT/src/test/java/com/stratio/crossdata/testsAT/specs/clusters/DetachConnector.java
@@ -1,7 +1,7 @@
 package com.stratio.crossdata.testsAT.specs.clusters;
 
-import com.stratio.crossdata.specs.BaseSpec;
-import com.stratio.crossdata.specs.Common;
+import com.stratio.crossdata.testsAT.specs.BaseSpec;
+import com.stratio.crossdata.testsAT.specs.Common;
 
 import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
