+++ testsAT/src/test/java/com/stratio/crossdata/testsAT/specs/exceptions/ConnectionExceptionSpec.java
@@ -1,7 +1,7 @@
-package com.stratio.crossdata.specs.exceptions;
+package com.stratio.crossdata.testsAT.specs.exceptions;
 
-import com.stratio.crossdata.specs.BaseSpec;
-import com.stratio.crossdata.specs.Common;
+import com.stratio.crossdata.testsAT.specs.BaseSpec;
+import com.stratio.crossdata.testsAT.specs.Common;
 import com.stratio.crossdata.common.exceptions.ConnectionException;
 import com.stratio.crossdata.common.exceptions.ExecutionException;
 import com.stratio.crossdata.common.exceptions.ParsingException;
