+++ testsAT/src/test/java/com/stratio/tests/utils/BaseTest.java
@@ -1,4 +1,4 @@
-package com.stratio.crossdata.testsAT.specs.utils;
+package com.stratio.tests.utils;
 
 import java.lang.reflect.Method;
