+++ testsAT/src/test/resources/log4j2.xml
@@ -1,17 +1,33 @@
 <?xml version="1.0" encoding="UTF-8"?>
+<!--
+    Copyright (C) 2015 Stratio (http://stratio.com)
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+            http://www.apache.org/licenses/LICENSE-2.0
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+-->
 <Configuration status="INFO">
+
     <Properties>
         <Property name="logLevel">INFO</Property>
     </Properties>
+
     <Appenders>
         <Console name="Console" target="SYSTEM_OUT">
-            <PatternLayout
-                    pattern="%highlight{%level{length=1} %12.12c{1}} %d{HH:mm:ss} (%16.16F:%3L) - %highlight{%m}%n" />
+            <PatternLayout pattern="%highlight{%level{length=1} %12.12c{1}} %d{HH:mm:ss} (%16.16F:%3L) - %highlight{%m}%n" />
         </Console>
     </Appenders>
     <Loggers>
+        <logger name="org.apache.spark" level="ERROR"/>
+        <logger name="com.stratio.crossdata.driver" level="ERROR"/>
+        <logger name="org.apache.flume.test.util.Agent" level="ERROR"/>
         <Root level="${sys:logLevel}">
-            <AppenderRef ref="Console" />
+            <AppenderRef ref="Console"/>
         </Root>
     </Loggers>
 </Configuration>
